pyproject.toml
[project]
name = "meo"
version = "0.1.0"
description = "Markdown Edit Orchestrator - TUI tool for structured markdown editing"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.40.0",
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "textual-dev>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
meo = "meo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meo"]

[tool.ruff]
line-length = 100
